+++ modules/tinymce/src/core/test/ts/browser/EditorSettingsTest.ts
@@ -306,25 +306,6 @@ describe('browser.tinymce.core.EditorSettingsTest', () => {
       );
     });
 
-    it('Merged settings when theme is mobile, forced_plugins in default override settings with user mobile settings (mobile)', () => {
-      assert.deepEqual(
-        EditorSettings.combineSettings(true, true, {}, { forced_plugins: [ 'a' ] }, {
-          plugins: [ 'b' ],
-          mobile: { plugins: [ 'lists custom' ], theme: 'mobile', toolbar_sticky: true }
-        }),
-        {
-          ...expectedPhoneDefaultSettings,
-          validate: true,
-          external_plugins: {},
-          forced_plugins: [ 'a' ],
-          plugins: 'a lists',
-          theme: 'mobile',
-          toolbar_sticky: true
-        },
-        'Should have forced_plugins merged with mobile plugins but only whitelisted user plugins'
-      );
-    });
-
     it('Merged settings when theme is silver, forced_plugins in default override settings with user mobile settings (mobile)', () => {
       assert.deepEqual(
         EditorSettings.combineSettings(true, true, {}, { forced_plugins: [ 'a' ] }, { plugins: [ 'b' ], mobile: { plugins: [ 'lists custom' ] }}),
@@ -341,27 +322,11 @@ describe('browser.tinymce.core.EditorSettingsTest', () => {
       );
     });
 
-    it('Merged settings when theme is mobile, on a mobile device', () => {
-      assert.deepEqual(
-        EditorSettings.combineSettings(true, true, {}, {}, { mobile: { theme: 'mobile' }, plugins: [ 'lists', 'b', 'autolink' ] }),
-        { ...expectedPhoneDefaultSettings, validate: true, external_plugins: {}, plugins: 'lists autolink', theme: 'mobile' },
-        'Should fallback to filtered white listed. settings.plugins'
-      );
-    });
-
     it('Merged settings when mobile.plugins is undefined, on a mobile device', () => {
       assert.deepEqual(
         EditorSettings.combineSettings(true, true, {}, {}, { theme: 'silver', plugins: [ 'lists', 'b', 'autolink' ], mobile: {}}),
         { ...expectedPhoneDefaultSettings, validate: true, external_plugins: {}, plugins: 'lists b autolink', theme: 'silver' },
         'Should use settings.plugins when mobile theme is not set'
-      );
-    });
-
-    it('Merged settings when mobile.plugins is undefined, legacy mobile theme and on a mobile device', () => {
-      assert.deepEqual(
-        EditorSettings.combineSettings(true, true, {}, {}, { theme: 'silver', plugins: [ 'lists', 'b', 'autolink' ], mobile: { theme: 'mobile' }}),
-        { ...expectedPhoneDefaultSettings, validate: true, external_plugins: {}, plugins: 'lists autolink', theme: 'mobile' },
-        'Should fallback to filtered white listed. settings.plugins'
       );
     });
 
@@ -378,14 +343,6 @@ describe('browser.tinymce.core.EditorSettingsTest', () => {
         EditorSettings.combineSettings(true, true, {}, {}, { mobile: { plugins: [ 'lists', 'autolink', 'foo', 'bar' ] }}),
         { ...expectedPhoneDefaultSettings, validate: true, external_plugins: {}, plugins: 'lists autolink foo bar' },
         'Should allow all plugins'
-      );
-    });
-
-    it('Merged settings with defined mobile.plugins and legacy mobile theme', () => {
-      assert.deepEqual(
-        EditorSettings.combineSettings(true, true, {}, {}, { mobile: { theme: 'mobile', plugins: [ 'lists', 'autolink', 'foo', 'bar' ] }}),
-        { ...expectedPhoneDefaultSettings, validate: true, external_plugins: {}, plugins: 'lists autolink', theme: 'mobile' },
-        'Should fallback to filtered white listed'
       );
     });