+++ modules/tinymce/src/themes/mobile/test/ts/module/test/ui/TestSelectors.ts
@@ -1,4 +1,0 @@
-import { Fun } from '@ephox/katamari';
-
-export const link = Fun.constant('.tinymce-mobile-toolbar-button');
-export const fontsize = Fun.constant('.tinymce-mobile-icon-font-size');